+++ packages/getbloc/lib/getbloc.dart
@@ -1,7 +1,8 @@
 library getbloc;
 
+export 'package:get/get.dart';
+
 export 'src/controller.dart';
 export 'src/listener_widget.dart';
 export 'src/observer_controller.dart';
-export 'src/test_controller.dart';
 export 'src/transform_controller.dart';