+++ packages/getbloc/test/controllers/controllers.dart
@@ -1,5 +1,5 @@
 export './async/async_controller.dart';
-export './complex/complex.dart';
+export './complex/complex_controller.dart';
 export './counter/counter.dart';
 export './seeded/seeded_controller.dart';
 export './simple/simple_controller.dart';