+++ packages/getbloc_test/lib/getbloc_test.dart
@@ -1,0 +1,5 @@
+library getbloc_test;
+
+export 'package:mocktail/mocktail.dart';
+
+export 'src/test_controller.dart';