+++ packages/getbloc_test/test/controllers/controllers.dart
@@ -1,0 +1,9 @@
+export 'async_counter_controller.dart';
+export 'complex_controller.dart';
+export 'counter_controller.dart';
+export 'debounce_counter_controller.dart';
+export 'error_counter_controller.dart';
+export 'exception_counter_controller.dart';
+export 'instant_emit_controller.dart';
+export 'multi_counter_controller.dart';
+export 'side_effect_counter_controller.dart';