+++ test/controllers/complex/complex.dart
@@ -1,2 +1,0 @@
-export './complex_controller.dart';
-export './complex_test_controller.dart';